+++ src/main/java/TaskList.java
@@ -1,0 +1,28 @@
+import java.util.ArrayList;
+
+public class TaskList {
+
+    private ArrayList<Task> taskList;
+
+    public TaskList() {
+        this.taskList = new ArrayList<>();
+    }
+
+    public int size() {
+        return this.taskList.size();
+    }
+
+    public Task get(int index) {
+        return this.taskList.get(index);
+    }
+
+    public void add(Task task) {
+        this.taskList.add(task);
+    }
+
+    public void remove(int index) {
+        this.taskList.remove(index);
+    }
+
+
+}
